Increased the size of the hash, preparing to make transposition table.

+++ src/zobrist.h
@@ -8,7 +8,7 @@
 #define ZOB_EN_PASSANT_START 772
 #define ZOB_TO_PLAY 780
 #define ZOB_TABLE_SIZE 781
-typedef uint32_t zob_hash_t;
+typedef uint64_t zob_hash_t;
 zob_hash_t* ZOBRIST_TABLE;
 void initZobrist();
 void closeZobrist();
